Tower targeting and shooting (closes #30)

Closes #30

See merge request szt-ab-2022/group-01/tree-o!26

+++ Assets/Scripts/Logic/Data/World/TilePosition.cs
@@ -23,8 +23,8 @@ public readonly struct TilePosition {
 		return (X - from.X) * (X - from.X) + (Y - from.Y) * (Y - from.Y);
 	}
 
-	public double Distance(TilePosition from) {
-		return Math.Sqrt(Distance2(from));
+	public float Distance(TilePosition from) {
+		return (float)Math.Sqrt(Distance2(from));
 	}
 
 	public override string ToString() {
